Create Bangla City
Virtual Space is not separate than the real space. Real people will join to the virtual space and whilst doing that they will bring their own baggages of language, culture and cultural preferences. Whilst our aim is to be part of global community, we cannot ignore the fact that most of people still are more comfortable to be with the people who share similar language, understanding of history and share their world view. 

Virtual world has both global and location specific interest. The Bangla City plans to bringing together who speaks in Bangla. There are around 300 million people globally who speaks the language but their presence in the virtual world is still in infancy. The Bangla City plans to use 3D space to bring together Bengalis/Bangladeshis living around the world to this place.

This space could be used as a meeting place, place of education in Bangla, a place of interaction for youth; a game zone and a fun zone. 

Though it will have aim to make this place primarily for Bengalis, but it won't restrict anyone who wishes to come, visit and 
participate to game, fun or development. Therefore, it will have both global and local interest to be fulfilled. 

As the aim is not to make this a boring place, but a place of entertainment therefore, there will be scope to monetise the space.